don't overwrite the file

+++ lib/hacer.rb
@@ -3,8 +3,10 @@ require 'yaml'
 module Hacer
   class Todolist
     def initialize(filename)
-      File.open(filename,'w') do |file|
-        YAML.dump([],file)
+      unless File.exists?(filename)
+        File.open(filename,'w') do |file|
+          YAML.dump([],file)
+        end
       end
     end
   end
